feat: add more entities to track emotions and more

+++ src/main/java/com/palmerodev/harmoni_api/model/request/SettingsRequest.java
@@ -2,5 +2,5 @@ package com.palmerodev.harmoni_api.model.request;
 
 import java.util.Map;
 
-public record SettingsRequest(Integer userId, Map<String, Object> settings) {
+public record SettingsRequest(Long userId, Map<String, Object> settings) {
 }
