+++ src/main/java/com/palmerodev/harmoni_api/model/request/UserInfoRequest.java
@@ -2,7 +2,7 @@ package com.palmerodev.harmoni_api.model.request;
 
 import com.palmerodev.harmoni_api.model.enums.Role;
 
-public record UserInfoRequest(Integer id,
+public record UserInfoRequest(Long id,
                               String name,
                               String email,
                               String password,
